17929 Add Scope Mixins to Prefix
